refactor: submitButton 관련 함수 이동
handleSubmit()과 그 안의 함수들을 AddItem에서 SubmitButton으로 이동
